refactor(ascend): consolidate Marlin MoE implementation in operator
