--- src/flaggems_vllm/runtime/backend/_ascend/ops/marlin_w4a16/.gitignore
@@ -1,2 +1,0 @@
-_build/
-__pycache__/
