--- src/flaggems_vllm/runtime/backend/_ascend/ops/marlin_w4a16/__init__.py
@@ -1,3 +1,0 @@
-# Copyright 2026 FlagOS Contributors
-# SPDX-License-Identifier: Apache-2.0
-"""TLE Marlin kernels using minimal FlagTree AscendC primitives."""
